Added a bunch of stuff
Fixed player movement
Added win/loss system
added areas where player must use abilities to navigate to reach orbs.
