+++ Assets/Scripts/Game/Objects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b3cf455f333ff744394953d4fdf47043
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
