+++ Assets/Scripts/Game/PlayerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd17c26afbeb9fb45a328d599e49fa5f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
