+++ Assets/Scripts/Game/WinLoss.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: daf1eb5115897ae4db7c99f120b8daf4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
