+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,4 +11,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/MainMenu.unity
     guid: d38c01736d8d6e342a7d2d7efda15e52
+  - enabled: 1
+    path: Assets/Scenes/CreditsScreen.unity
+    guid: 342efb603f925a243b1332b30a86e668
   m_configObjects: {}
